refactor to new directory partition
