+++ internal/partition/log_writer.go
@@ -1,4 +1,4 @@
-package replication
+package partition
 
 import (
 	"fmt"
